fix: address Copilot review — deduplicate theme helpers and fix render-time mutation
Copilot review comment 1 (doenetml-iframe/src/index.tsx):
  The file duplicated ThemeSetting/ResolvedTheme types, getSystemTheme,
  subscribeToSystemTheme, and useResolvedTheme from doenetml/src/utils/theme.ts.
  Fix: extract the pure (non-React) theme helpers — ThemeSetting, ResolvedTheme,
  getSystemTheme, subscribeToSystemTheme — into a new packages/utils/src/theme/theme.ts
  and re-export from @doenet/utils. Both doenetml/src/utils/theme.ts and
  doenetml-iframe/src/index.tsx now import these shared helpers from @doenet/utils
  rather than duplicating them. The React hook (useResolvedTheme) remains local
  to each package since @doenet/utils has no React dependency.

Copilot review comment 2 (unique-keyboard-tray.tsx):
  virtualKeyboardState.nextRegistrationId was mutated during render to allocate
  the registration ID, which is unsafe in React 18 concurrent mode and StrictMode
  double-invocation. Fix: use a lazy useState initializer
  (React.useState<number>(() => { ... })) which runs exactly once at component
  mount (commit time) rather than during render.
