+++ packages/utils/src/index.ts
@@ -20,6 +20,7 @@ export * from "./math/subset-of-reals-operations";
 export * from "./style/colorWords";
 export * from "./style/style";
 export * from "./style/styleDefinitionHelpers";
+export * from "./theme/theme";
 export * from "./types/core";
 export * from "./url/url";
 export * from "./version/doenetMLversion";
